+++ ios/Classes/TiAdmobTypes.h
@@ -9,5 +9,6 @@
 
 NS_ENUM(NSUInteger, TiAdmobAdType){
   TiAdmobAdTypeBanner = 0,
-  TiAdmobAdTypeInterstitial
+  TiAdmobAdTypeInterstitial,
+  TiAdmobAdTypeRewardedVideo
 };