+++ ios/Classes/TiAdmobViewProxy.m
@@ -6,6 +6,8 @@
  */
 
 #import "TiAdmobViewProxy.h"
+#import "TiAdmobTypes.h"
+#import "TiApp.h"
 #import "TiUtils.h"
 
 @implementation TiAdmobViewProxy
@@ -18,6 +20,23 @@
 - (void)receive:(id)unused
 {
   [[self admobView] initialize];
+}
+
+- (void)showRewardedVideo:(id)args {
+  if ([[GADRewardBasedVideoAd sharedInstance] isReady]) {
+    [GADRewardBasedVideoAd.sharedInstance presentFromRootViewController:TiApp.app.controller.topPresentedController];
+  }
+}
+
+- (void)loadRewardedVideo:(id)adUnitId
+{
+  id adType = [self valueForKey:@"adType"];
+  if (adType != nil && [TiUtils boolValue:adType def:TiAdmobAdTypeBanner] != TiAdmobAdTypeRewardedVideo) {
+    return;
+  }
+
+  ENSURE_SINGLE_ARG(adUnitId, NSString);
+  [GADRewardBasedVideoAd.sharedInstance loadRequest:self.admobView.request withAdUnitID:adUnitId];
 }
 
 @end