+++ ios/platform/GoogleMobileAds.framework/Headers/GADCustomEventParameters.h
@@ -2,18 +2,17 @@
 //  GADCustomEventParameters.h
 //  Google Mobile Ads SDK
 //
-//  Copyright 2016 Google Inc. All rights reserved.
+//  Copyright 2016 Google LLC. All rights reserved.
 //
 
 #import <Foundation/Foundation.h>
-
 #import <GoogleMobileAds/GoogleMobileAdsDefines.h>
 
-GAD_ASSUME_NONNULL_BEGIN
+NS_ASSUME_NONNULL_BEGIN
 
 /// Key for getting the server parameter configured in AdMob when mediating to a custom event
 /// adapter.
 /// Example: NSString *serverParameter = connector.credentials[GADCustomEventParametersServer].
 GAD_EXTERN NSString *const GADCustomEventParametersServer;
 
-GAD_ASSUME_NONNULL_END
+NS_ASSUME_NONNULL_END