+++ ios/platform/GoogleMobileAds.framework/Headers/GADNativeAdViewAdOptions.h
@@ -2,13 +2,13 @@
 //  GADNativeAdViewAdOptions.h
 //  Google Mobile Ads SDK
 //
-//  Copyright 2016 Google Inc. All rights reserved.
+//  Copyright 2016 Google LLC. All rights reserved.
 //
 
 #import <GoogleMobileAds/GADAdLoader.h>
 #import <GoogleMobileAds/GoogleMobileAdsDefines.h>
 
-GAD_ASSUME_NONNULL_BEGIN
+NS_ASSUME_NONNULL_BEGIN
 
 /// Position of the AdChoices icon in the containing ad.
 typedef NS_ENUM(NSInteger, GADAdChoicesPosition) {
@@ -26,4 +26,4 @@ typedef NS_ENUM(NSInteger, GADAdChoicesPosition) {
 
 @end
 
-GAD_ASSUME_NONNULL_END
+NS_ASSUME_NONNULL_END