+++ ios/platform/GoogleMobileAds.framework/Headers/GADUnifiedNativeAd.h
@@ -2,22 +2,23 @@
 //  GADUnifiedNativeAd.h
 //  Google Mobile Ads SDK
 //
-//  Copyright 2017 Google Inc. All rights reserved.
+//  Copyright 2017 Google LLC. All rights reserved.
 //
 
 #import <Foundation/Foundation.h>
-#import <UIKit/UIKit.h>
-
 #import <GoogleMobileAds/GADAdChoicesView.h>
 #import <GoogleMobileAds/GADAdLoaderDelegate.h>
+#import <GoogleMobileAds/GADMediaContent.h>
 #import <GoogleMobileAds/GADMediaView.h>
+#import <GoogleMobileAds/GADMuteThisAdReason.h>
 #import <GoogleMobileAds/GADNativeAdImage.h>
-#import <GoogleMobileAds/GADVideoController.h>
 #import <GoogleMobileAds/GADUnifiedNativeAdAssetIdentifiers.h>
 #import <GoogleMobileAds/GADUnifiedNativeAdDelegate.h>
+#import <GoogleMobileAds/GADVideoController.h>
 #import <GoogleMobileAds/GoogleMobileAdsDefines.h>
+#import <UIKit/UIKit.h>
 
-GAD_ASSUME_NONNULL_BEGIN
+NS_ASSUME_NONNULL_BEGIN
 
 /// Unified native ad. To request this ad type, pass kGADAdLoaderAdTypeUnifiedNative
 /// (see GADAdLoaderAdTypes.h) to the |adTypes| parameter in GADAdLoader's initializer method. If
@@ -48,13 +49,13 @@ GAD_ASSUME_NONNULL_BEGIN
 @property(nonatomic, readonly, copy, nullable) NSString *price;
 /// Identifies the advertiser. For example, the advertiser’s name or visible URL.
 @property(nonatomic, readonly, copy, nullable) NSString *advertiser;
-/// Video controller for controlling video playback in GADUnifiedNativeAdView's mediaView.
-@property(nonatomic, strong, readonly, nullable) GADVideoController *videoController;
 
 /// Optional delegate to receive state change notifications.
 @property(nonatomic, weak, nullable) id<GADUnifiedNativeAdDelegate> delegate;
 
-/// Root view controller for handling ad actions.
+/// Reference to a root view controller that is used by the ad to present full screen content after
+/// the user interacts with the ad. The root view controller is most commonly the view controller
+/// displaying the ad.
 @property(nonatomic, weak, nullable) UIViewController *rootViewController;
 
 /// Dictionary of assets which aren't processed by the receiver.
@@ -64,6 +65,17 @@ GAD_ASSUME_NONNULL_BEGIN
 /// AdMob ads, this method returns @"GADMAdapterGoogleAdMobAds". For ads fetched via mediation
 /// custom events, this method returns @"GADMAdapterCustomEvents".
 @property(nonatomic, readonly, copy, nullable) NSString *adNetworkClassName;
+
+/// Indicates whether custom Mute This Ad is available for the native ad.
+@property(nonatomic, readonly, getter=isCustomMuteThisAdAvailable) BOOL customMuteThisAdAvailable;
+
+/// An array of Mute This Ad reasons used to render customized mute ad survey. Use this array to
+/// implement your own Mute This Ad feature only when customMuteThisAdAvailable is YES.
+@property(nonatomic, readonly, nullable) NSArray<GADMuteThisAdReason *> *muteThisAdReasons;
+
+/// Media content. Set the associated media view's mediaContent property to this object to display
+/// this content.
+@property(nonatomic, readonly, nonnull) GADMediaContent *mediaContent;
 
 /// Registers ad view, clickable asset views, and nonclickable asset views with this native ad.
 /// Media view shouldn't be registered as clickable.
@@ -80,12 +92,23 @@ GAD_ASSUME_NONNULL_BEGIN
 /// unregistered.
 - (void)unregisterAdView;
 
+/// Reports the mute event with the mute reason selected by user. Use nil if no reason was selected.
+/// Call this method only if customMuteThisAdAvailable is YES.
+- (void)muteThisAdWithReason:(nullable GADMuteThisAdReason *)reason;
+
+#pragma mark - Deprecated
+
+/// Video controller for controlling video playback in GADUnifiedNativeAdView's mediaView.
+@property(nonatomic, readonly, nullable)
+    GADVideoController *videoController GAD_DEPRECATED_MSG_ATTRIBUTE(
+        "Use the videoController property from the ad's mediaContent instead.");
+
 @end
 
 #pragma mark - Protocol and constants
 
 /// The delegate of a GADAdLoader object implements this protocol to receive GADUnifiedNativeAd ads.
-@protocol GADUnifiedNativeAdLoaderDelegate<GADAdLoaderDelegate>
+@protocol GADUnifiedNativeAdLoaderDelegate <GADAdLoaderDelegate>
 /// Called when a unified native ad is received.
 - (void)adLoader:(GADAdLoader *)adLoader didReceiveUnifiedNativeAd:(GADUnifiedNativeAd *)nativeAd;
 @end
@@ -120,10 +143,10 @@ GAD_ASSUME_NONNULL_BEGIN
 /// Weak reference to your ad view's media asset view.
 @property(nonatomic, weak, nullable) IBOutlet GADMediaView *mediaView;
 /// Weak reference to your ad view's AdChoices view. Must set adChoicesView before setting
-/// nativeAd, otherwise AdChoices will be rendered in the publisher's preferredAdChoicesPosition as
+/// nativeAd, otherwise AdChoices will be rendered according to the preferredAdChoicesPosition
 /// defined in GADNativeAdViewAdOptions.
 @property(nonatomic, weak, nullable) IBOutlet GADAdChoicesView *adChoicesView;
 
 @end
 
-GAD_ASSUME_NONNULL_END
+NS_ASSUME_NONNULL_END